分析NameServer启动原理——源码分析

+++ store/src/main/java/org/apache/rocketmq/store/config/MessageStoreConfig.java
@@ -200,7 +200,7 @@ public class MessageStoreConfig {
     public int getMappedFileSizeConsumeQueue() {
 
         int factor = (int) Math.ceil(this.mappedFileSizeConsumeQueue / (ConsumeQueue.CQ_STORE_UNIT_SIZE * 1.0));
-        return (int) (factor * ConsumeQueue.CQ_STORE_UNIT_SIZE);
+        return factor * ConsumeQueue.CQ_STORE_UNIT_SIZE;
     }
 
     public void setMappedFileSizeConsumeQueue(int mappedFileSizeConsumeQueue) {
